Undo a premature refactoring.
...because you know what they say, "A premature refactoring is the hobgoblin of
little minds, adored by little programmers and code-monkeys and
pythonistas."
